requirements.txt
numpy==1.22.4
opencv-python==4.5.5.64
Pillow==9.1.1
PyAutoGUI==0.9.53
PyQt5==5.15.4
ruamel.yaml==0.17.21